InnerI-Secure-Repo/gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
jsonschema==4.23.0
cryptography==43.0.3
python-dotenv==1.0.1
PyJWT==2.9.0
